Moved dropUnusedFiles()
